Vulkan 1.4 : Use local-only memory for vertex reading (#39)
Do not use a host consistent buffer for the vertex buffer, to avoid latency issues with the device.
NB: Ideally speaking we should be using an allocator that manages one allocation for multiple buffers as the max number of concurrent VkAllocateMemory calls can be as low as 4096. So the ideal solution would be an allocator layer which does one VkAllocateMemory at max possible size and then indexes through the allocated memory to retrieve the needed buffer.